Added 2023 day 5

+++ 2023/newday.ps1
@@ -1,5 +1,8 @@
 param (
     [int] $Day
 )
-[System.IO.File]::WriteAllText("$($PSScriptRoot)\AoC2023\Days\Day$($Day).cs",[System.IO.File]::ReadAllText("$($PSScriptRoot)\AoC2023\Days\DayTemplate.cs").Replace("Template", $day, [System.StringComparison]::Ordinal))
-[System.IO.File]::WriteAllText("$($PSScriptRoot)\AoC2023.UnitTests\DayTests\Day$($Day)Tests.cs",[System.IO.File]::ReadAllText("$($PSScriptRoot)\AoC2023.UnitTests\DayTests\DayTemplateTests.cs").Replace("Template", $day, [System.StringComparison]::Ordinal))
+if (-not (Test-Path "$($PSScriptRoot)\AoC2023\Days\Day$($Day).cs"))
+{
+	[System.IO.File]::WriteAllText("$($PSScriptRoot)\AoC2023\Days\Day$($Day).cs",[System.IO.File]::ReadAllText("$($PSScriptRoot)\AoC2023\Days\DayTemplate.cs").Replace("Template", $day, [System.StringComparison]::Ordinal))
+	[System.IO.File]::WriteAllText("$($PSScriptRoot)\AoC2023.UnitTests\DayTests\Day$($Day)Tests.cs",[System.IO.File]::ReadAllText("$($PSScriptRoot)\AoC2023.UnitTests\DayTests\DayTemplateTests.cs").Replace("Template", $day, [System.StringComparison]::Ordinal))
+}
